move ban() to vmod std
