+++ bin/varnishd/cache/cache_vrt.c
@@ -402,78 +402,6 @@ VRT_synth_page(VRT_CTX, const char *str, ...)
 	va_end(ap);
 }
 
-/*--------------------------------------------------------------------*/
-
-void
-VRT_ban_string(VRT_CTX, const char *str)
-{
-	char *a1, *a2, *a3;
-	char **av;
-	struct ban_proto *bp;
-	const char *err;
-	int i;
-
-	CHECK_OBJ_NOTNULL(ctx, VRT_CTX_MAGIC);
-	AN(ctx->vsl);
-	AN(str);
-
-	bp = BAN_Build();
-	if (bp == NULL) {
-		VSLb(ctx->vsl, SLT_VCL_Error, "ban(): Out of Memory");
-		return;
-	}
-	av = VAV_Parse(str, NULL, ARGV_NOESC);
-	AN(av);
-	if (av[0] != NULL) {
-		VSLb(ctx->vsl, SLT_VCL_Error, "ban(): %s", av[0]);
-		VAV_Free(av);
-		BAN_Abandon(bp);
-		return;
-	}
-	for (i = 0; ;) {
-		a1 = av[++i];
-		if (a1 == NULL) {
-			VSLb(ctx->vsl, SLT_VCL_Error,
-			    "ban(): No ban conditions found.");
-			break;
-		}
-		a2 = av[++i];
-		if (a2 == NULL) {
-			VSLb(ctx->vsl, SLT_VCL_Error,
-			    "ban(): Expected comparison operator.");
-			break;
-		}
-		a3 = av[++i];
-		if (a3 == NULL) {
-			VSLb(ctx->vsl, SLT_VCL_Error,
-			    "ban(): Expected second operand.");
-			break;
-		}
-		err = BAN_AddTest(bp, a1, a2, a3);
-		if (err) {
-			VSLb(ctx->vsl, SLT_VCL_Error, "ban(): %s", err);
-			break;
-		}
-		if (av[++i] == NULL) {
-			err = BAN_Commit(bp);
-			if (err == NULL)
-				bp = NULL;
-			else
-				VSLb(ctx->vsl, SLT_VCL_Error, "ban(): %s", err);
-			break;
-		}
-		if (strcmp(av[i], "&&")) {
-			VSLb(ctx->vsl, SLT_VCL_Error,
-			    "ban(): Expected && between conditions,"
-			    " found \"%s\"", av[i]);
-			break;
-		}
-	}
-	if (bp != NULL)
-		BAN_Abandon(bp);
-	VAV_Free(av);
-}
-
 VCL_BYTES
 VRT_CacheReqBody(VRT_CTX, VCL_BYTES maxsize)
 {
